+++ src/main/java/C13Group2/BankingAPI/enums/BillStatus.java
@@ -1,4 +1,4 @@
-package C13Group2.BankingAPI;
+package C13Group2.BankingAPI.enums;
 
 public enum BillStatus {
     Pending("pending"),
