+++ src/main/java/C13Group2/BankingAPI/response/ErrorResponse.java
@@ -1,4 +1,4 @@
-package C13Group2.BankingAPI;
+package C13Group2.BankingAPI.response;
 
 public class ErrorResponse {
